make_class_heritate
