--- articleapp/class/JavaProject_class.py
@@ -1,6 +1,0 @@
-from ABC import Project
-
-
-class PythonProject(Project):
-    def __init__(self):
-        super(PythonProject, self).__init__()
