+++ articleapp/classes/ABC/__init__.py
@@ -2,3 +2,4 @@ from Article_class import Article
 from Project_class import Project
 from Category_class import Category
 from User_class import User
+from Team_class import Team
