+++ articleapp/classes/ProjectCategory.py
@@ -1,0 +1,26 @@
+from articleapp.classes.ABC import Category
+
+
+class ProjectCategory(Category):
+
+    def __init__(self):
+        super(Category, self).__init__()
+
+    def make(self):
+        pass
+
+    def search_title(self):
+        pass
+
+    def search_project(self):
+        pass
+
+    def search_writer(self):
+        pass
+
+    def pagination(self):
+        pass
+
+
+
+
