+++ articleapp/service/project_service.py
@@ -1,2 +1,4 @@
 import typing
+from articleapp.classes.ABC import Article, Category, Project, Team, User
+
