+++ articleapp/test/service/test_project_service.py
@@ -1,5 +1,4 @@
 from django.test.testcases import TestCase
-from articleapp.service.project_service import ProjectArticle
 
 
 class ProjectArticleClassTest(TestCase):
